+++ package.json
@@ -16,6 +16,7 @@
 		"@types/react": "18.2.38",
 		"@types/react-dom": "18.2.17",
 		"@vercel/analytics": "^1.1.1",
+		"@vercel/speed-insights": "^1.0.2",
 		"autoprefixer": "10.4.16",
 		"eslint-config-next": "^14.0.3",
 		"framer-motion": "^10.16.5",
