+++ src/app/(Recipes pages)/categorias/[id]/page.tsx
@@ -1,7 +1,7 @@
-import CardRecipe from '@/SharedComponents/CardRecipe'
-import DynamicTitle from '@/SharedComponents/DynamicTitle'
-import NotFound from '@/SharedComponents/NotFound'
-import ShowOneRecipe from '@/SharedComponents/ShowOneRecipe'
+import CardRecipe from '@/components/shared/CardRecipe'
+import DynamicTitle from '@/components/shared/DynamicTitle'
+import NotFound from '@/components/shared/NotFound'
+import ShowOneRecipe from '@/components/shared/ShowOneRecipe'
 import { type ParamProps } from '@interfaces/ParamProps.interface'
 import { Metadata } from 'next'
 import { notFound } from 'next/navigation'
