+++ src/app/(Recipes pages)/paises/[id]/page.tsx
@@ -1,8 +1,8 @@
 import { GetAllRecipesOneCountry } from '@/(Recipes pages)/functions/GetAllRecipesOneCountry'
-import CardRecipe from '@/SharedComponents/CardRecipe'
-import DynamicTitle from '@/SharedComponents/DynamicTitle'
-import NotFound from '@/SharedComponents/NotFound'
-import ShowOneRecipe from '@/SharedComponents/ShowOneRecipe'
+import CardRecipe from '@/components/shared/CardRecipe'
+import DynamicTitle from '@/components/shared/DynamicTitle'
+import NotFound from '@/components/shared/NotFound'
+import ShowOneRecipe from '@/components/shared/ShowOneRecipe'
 import { type ParamProps } from '@/interfaces/ParamProps.interface'
 import { Metadata } from 'next'
 import { notFound } from 'next/navigation'
