+++ src/app/(Recipes pages)/recetas/components/allRecipes/SkeletonRecipe.tsx
@@ -1,4 +1,4 @@
-import SkeletonCard from '@/SharedComponents/skeletons/SkeletonCardRecipe'
+import SkeletonCard from '@/components/skeletons/SkeletonCardRecipe'
 
 const SkeletonRecipe = () => {
 	return (
