+++ src/app/components/ui/index.ts
@@ -1,0 +1,4 @@
+import UiImage from './Image'
+import Logo from './Logo'
+import UiPagination from './Pagination'
+export { Logo, UiImage, UiPagination }
