+++ src/app/lib/ResponsiveCaruselOptions.ts
